SmartHealthChain_AIPart/requirements.txt
mysql-connector-python==8.0.33 
pandas==2.0.3 
scikit-learn==1.3.0 
transformers==4.30.2 
torch==2.0.1 
numpy==1.24.3 
flask==2.3.2